api-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-auth==2.23.0
pydantic==2.4.0
pydantic-settings==2.0.3
